+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+In Development
+--------------
+
+* The old `email` field on the returned `ValidatedEmail` object, which in the previous version was superseded by `normalized`, will now raise a deprecation warning if used. See https://stackoverflow.com/q/879173 for strategies to suppress the DeprecationWarning.
+* A `__version__` module attribute is added.
+
 2.0.0 (April 15, 2023)
 ----------------------